requirements.txt
openai>=1.32.0
python-telegram-bot>=20.5
